** Added rest endpoint for reindexing Hibernate entities
