+++ webapp/src/main/java/fi/foyt/fni/rest/illusion/OAuthScopes.java
@@ -28,6 +28,7 @@ public enum OAuthScopes {
   USERS_USER_FIND_ME, 
   
   SYSTEM_JPA_CACHE_FLUSH, 
+  SYSTEM_REINDEX_SEARCH, 
   
   ILLUSION_CREATE_FORUM_POST, 
   ILLUSION_FIND_FORUM_POST,
